+++ QA/index.mjs
@@ -1,9 +1,6 @@
 // This script performs QA checks.
 
-import child_process from "child_process"
-import crypto from "crypto"
-import fs from "fs"
-import path from "path"
+import { runTestCase } from "./lib/runner.mjs"
 
 // hijackPopularDNSServers returns an object containing the rules
 // for hijacking popular DNS servers with `miniooni --censor`.
@@ -381,71 +378,6 @@ const testCases = [{
 
 ]
 
-// tempFile returns the name of a temporary file. This function is
-// not as secure as using mktemp but it does not matter in this
-// context. We just need to create file names in the local directory
-// with enough entropy that every run has a different name.
-//
-// See https://stackoverflow.com/questions/7055061 for an insightful
-// discussion on how one should or should not create temp files.
-function tempFile(suffix) {
-    return path.join(`tmp-${crypto.randomBytes(16).toString('hex')}.${suffix}`)
-}
-
-// exec executes a command. This function throws on failure.
-function exec(command) {
-    console.log(`+ ${command}`)
-    child_process.execSync(command)
-}
-
-// writeCensorJsonFile writes a censor.json file using a file name
-// containing random characters and returns the file name.
-function writeCensorJsonFile(testCase) {
-    const fileName = tempFile("json")
-    fs.writeFileSync(fileName, JSON.stringify(testCase.blocking))
-    return fileName
-}
-
-// readReportFile reads and parses the report file, thus returning
-// the JSON object contained inside the report file.
-function readReportFile(reportFile) {
-    const data = fs.readFileSync(reportFile, { "encoding": "utf-8" })
-    return JSON.parse(data)
-}
-
-// runExperiment runs the given test case with the given experiment.
-function runExperiment(testCase, experiment, checker) {
-    console.log(`## running: ${testCase.name}.${experiment}`)
-    console.log("")
-    const censorJson = writeCensorJsonFile(testCase)
-    const reportJson = tempFile("json")
-    exec(`./miniooni -n --censor ${censorJson} -o ${reportJson} -i ${testCase.input} ${experiment}`)
-    console.log("")
-    const report = readReportFile(reportJson)
-    const analysisResult = checker(experiment, report)
-    console.log("")
-    console.log("")
-    switch (analysisResult) {
-        case true:
-        case false:
-            return analysisResult
-        default:
-            console.log("the analysis function returned neither true nor false")
-            process.exit(1)
-    }
-}
-
-// runTestCase runs the given test case.
-function runTestCase(testCase) {
-    console.log("")
-    console.log(`# running: ${testCase.name}`)
-    let result = true
-    for (const [name, checker] of Object.entries(testCase.experiments)) {
-        result = result && runExperiment(testCase, name, checker)
-    }
-    return result
-}
-
 // checkForDuplicateTestCases ensures there are no duplicate names.
 function checkForDuplicateTestCases() {
     let dups = {}
@@ -479,14 +411,8 @@ function makeTestCasesMap() {
     return map
 }
 
-// recompileMiniooni recompiles miniooni if needed.
-function recompileMiniooni() {
-    exec("go build -v ./internal/cmd/miniooni")
-}
-
 // commandRun implements the run command.
 function commandRun(args) {
-    recompileMiniooni()
     const bailOnFailure = (result) => {
         if (!result) {
             console.log("some checks failed (see above logs)")
